component for major item table
