+++ apps/web-next/app/stats/majors.module.css
@@ -1,0 +1,30 @@
+.thin_border {
+  border-style: solid;
+  border-width: 1px;
+  border-color: #dddddd;
+  padding: 5px 5px 5px 5px;
+  margin: 0px;
+  border-collapse: collapse;
+}
+
+.legacy_style {
+  font-family: Arial, Helvetica, sans-serif;
+  background-color: white;
+  color: black;
+  font-size: 0.75em;
+  width: 100%;
+  border-spacing: 0;
+  border-collapse: collapse;
+}
+
+tr.header > th {
+  font-weight: bold;
+}
+
+.tan_cell {
+  background-color: tan;
+}
+
+.gray_cell {
+  background-color: lightgrey;
+}
